+++ app/assets/stylesheets/application.scss
@@ -48,3 +48,7 @@
 .solicitor_results_list {
   padding-left: 14px;
 }
+td.actions{
+  text-align: center;
+}
+
