+++ app/views/defence_requests/_form.html.erb
@@ -1,0 +1,119 @@
+<% if @defence_request.new_record? %>
+  <%= render partial: 'solicitor_search', locals: {object: @defence_request} %>
+<% end %>
+
+<%= bootstrap_form_for(@defence_request) do |f| %>
+  <%= render partial: 'shared/form_errors', locals: {object: @defence_request} %>
+
+  <% if @defence_request.new_record? %>
+    <%= f.form_group :solicitor_type, label: {text: t('solicitor_type')} do %>
+      <%= f.radio_button :solicitor_type, :duty, label: t('duty'), name: 'defence_request[solicitor_type]', inline: true, class: 'solicitor_type_radio', checked: true %>
+      <%= f.radio_button :solicitor_type, :own, label: t('own'), name: 'defence_request[solicitor_type]', inline: true, class: 'solicitor_type_radio' %>
+    <% end %>
+  <% end %>
+
+  <!-- details -->
+  <div class="panel panel-default details">
+    <div class="panel-heading">
+      <h3 class="panel-title"><%= t('details') %></h3>
+    </div>
+    <div class="panel-body">
+      <%= f.text_field :solicitor_name, label: t('solicitor_name'), autofocus: true %>
+      <%= f.text_field :solicitor_firm, label: t('solicitor_firm') %>
+      <%= f.select :scheme, solicitor_schemes, {label: t('scheme')} %>
+      <%= f.text_field :phone_number, label: t('phone_number') %>
+      <%= f.text_field :custody_number, label: t('custody_number') %>
+      <%= f.text_field :allegations, label: t('allegations') %>
+      <%= f.time_select :time_of_arrival, {label: t('time_of_arrival')}, {class: 'time-select'} %>
+    </div>
+  </div>
+  <!-- detainee -->
+  <div class="panel panel-default detainee">
+    <div class="panel-heading">
+      <h3 class="panel-title"><%= t('detainee') %></h3>
+    </div>
+    <div class="panel-body">
+      <%= f.text_field :detainee_surname, label: t('detainee_surname') %>
+      <%= f.text_field :detainee_first_name, label: t('detainee_first_name') %>
+
+      <%= f.form_group :gender, label: {text: t('gender')} do %>
+        <%= f.radio_button :gender, :male, label: t('male'), name: 'defence_request[gender]', inline: true %>
+        <%= f.radio_button :gender, :female, label: t('female'), name: 'defence_request[gender]', inline: true %>
+      <% end %>
+
+      <%= f.check_box :adult, label: t('adult') %>
+
+      <%= f.date_select :date_of_birth, {
+          start_year: Date.today.year,
+          end_year: 120.years.ago.year,
+          label: t('date_of_birth')
+      }, {
+                            class: 'time-select'
+                        } %>
+      <%= f.check_box :appropriate_adult, label: t('appropriate_adult') %>
+    </div>
+  </div>
+  <!-- comments -->
+  <div class="panel panel-default comments">
+    <div class="panel-heading">
+      <h3 class="panel-title"><%= t('comments') %></h3>
+    </div>
+    <div class="panel-body">
+      <%= f.text_area :comments, label: t('comments'), rows: '5' %>
+    </div>
+  </div>
+
+  <%= f.submit class: 'btn btn-primary' %>
+
+  <% if @defence_request.new_record? %>
+    <%= link_to t('cancel'), defence_requests_path, class: 'btn btn-default cancel' %>
+  <% else %>
+    <%= link_to t('delete'), defence_request_path(@defence_request), method: :delete, data: {confirm: t('confirm')}, class: 'btn btn-danger' %>
+    <%= link_to t('cancel'), defence_requests_path, class: 'btn btn-default cancel' %>
+  <% end %>
+
+<% end %>
+
+<% if @defence_request.new_record? %>
+  <script>
+    $(document).ready(function () {
+      var ownSolicitor_checked = ($("input[type='radio']#defence_request_solicitor_type_own:checked").length > 0);
+
+      var own_setup = function () {
+        $('#defence_request_solicitor_firm').removeAttr("disabled");
+        $('#defence_request_solicitor_name').removeAttr("disabled");
+        $('#defence_request_scheme').val("No Scheme");
+        $('#defence_request_scheme').attr("disabled", "disabled");
+      }
+
+      var duty_setup = function () {
+        $('#defence_request_solicitor_firm').attr("disabled", "disabled");
+        $('#defence_request_solicitor_firm').val("");
+        $('#defence_request_solicitor_name').attr("disabled", "disabled");
+        $('#defence_request_solicitor_name').val("");
+        $('#defence_request_scheme').removeAttr("disabled");
+        $('.solicitor_results_list').html("")
+        $('#q').val("")
+      }
+
+      if (ownSolicitor_checked) {
+        $('.search_box').show();
+        own_setup()
+      } else {
+        $('.search_box').hide();
+        duty_setup()
+      }
+
+      $(".solicitor_type_radio").change(function () {
+        if (this.value == "own") {
+          $('.search_box').stop(true, true).show(1000);
+          own_setup()
+        } else {
+          $('.search_box').stop(true, true).hide(1000);
+          duty_setup()
+        }
+      });
+
+    });
+  </script>
+<% end %>
