+++ app/views/defence_requests/edit.html.erb
@@ -1,0 +1,1 @@
+<%= render partial: 'shared/form_template', locals: { form_type: 'edit', object_name: 'defence_request' } %>
