+++ app/views/defence_requests/index.html.erb
@@ -28,6 +28,7 @@
       <th><%= t('allegations') %></th>
       <th><%= t('comments') %></th>
       <th><%= t('time_of_arrival') %></th>
+      <th><%= t('actions') %></th>
     </tr>
     </thead>
     <tbody>
