+++ app/views/shared/_form_template.html.erb
@@ -1,0 +1,12 @@
+<div class="col-md-12">
+  <div class="row">
+    <div class="panel panel-default col-md-8 col-md-offset-2">
+      <div class="panel-heading">
+        <h3 class="panel-title"><%= (form_type +' '+ object_name).humanize %></h3>
+      </div>
+      <div class="panel-body">
+        <%= render partial: 'form', locals: { form_type: form_type } %>
+      </div>
+    </div>
+  </div>
+</div>
