+++ config/locales/en.yml
@@ -1,60 +1,51 @@
-# Files in the config/locales directory are used for internationalization
-# and are automatically loaded by Rails. If you want to use locales other
-# than English, add the necessary files in this directory.
-#
-# To use the locales, use `I18n.t`:
-#
-#     I18n.t 'hello'
-#
-# In views, this is aliased to just `t`:
-#
-#     <%= t('hello') %>
-#
-# To use a different locale, set it with `I18n.locale`:
-#
-#     I18n.locale = :es
-#
-# This would use the information in config/locales/es.yml.
-#
-# To learn more, please read the Rails Internationalization guide
-# available at http://guides.rubyonrails.org/i18n.html.
-
 en:
-  not_authorized: "You are not authorised to perform this action"
-  new_defence_request: "New Defence Request"
-  prohibit_save: Please review the following %{count}
-  error: error
-  solicitor_type: "Solicitor Type"
-  search: "Search"
-  solicitor_name: "Solicitor Name"
-  solicitor_firm: "Solicitor Firm"
-  scheme: "Scheme"
-  phone_number: "Phone Number"
-  custody_number: "Custody Number"
-  allegations: "Allegations"
-  time_of_arrival: "Time of Arrival"
-  detainee_name: "Detainee Surname"
-  detainee_first_name: "Detainee First Name"
-  gender: "Gender"
-  male: "Male"
-  female: "Female"
+  actions: "Actions"
   adult: "Adult?"
-  date_of_birth: "Date of Birth"
+  allegations: "Allegations"
   appropriate_adult: "Appropriate Adult"
-  detainee_surname: "Detainee Surname"
+  cancel: "Cancel"
+  change_password: "Change Password"
   comments: "Comments"
-  enter_solicitor_firm_name: "Enter solicitor/firm name..."
+  confirm: "Confirm"
+  custody_number: "Custody Number"
+  dashboard: 'Dashboard'
+  date_of_birth: "Date of Birth"
+  delete: "Delete"
   details: "Details"
   detainee: "Detainee"
-  new_defence_request: "New Defence Request"
-  change_password: "Change Password"
-  own: 'Own'
+  detainee_first_name: "Detainee First Name"
+  detainee_name: "Detainee Surname"
+  detainee_surname: "Detainee Surname"
   duty: 'Duty'
+  enter_solicitor_firm_name: "Enter solicitor/firm name..."
+  error: error
+  female: "Female"
+  gender: "Gender"
+  male: "Male"
+  new_defence_request: "New Defence Request"
+  new_defence_request: "New Defence Request"
   no_results_found: 'No results found - Contact DSCC'
-  dashboard: 'Dashboard'
+  not_authorized: "You are not authorised to perform this action"
+  own: 'Own'
+  phone_number: "Phone Number"
+  prohibit_save: Please review the following %{count}
+  scheme: "Scheme"
+  search: "Search"
+  solicitor_firm: "Solicitor Firm"
+  solicitor_name: "Solicitor Name"
+  solicitor_type: "Solicitor Type"
+  time_of_arrival: "Time of Arrival"
+  update: "Update"
 
   models:
     create: "%{model} successfully created"
     update: "%{model} successfully updated"
     destroy: "%{model} successfully deleted"
     failed_destroy: "%{model} was not deleted"
+
+  helpers:
+    submit:
+      create: "Create %{model}"
+      update: "Update %{model}"
+
+
