+++ spec/policies/defence_requests_controller_policy_spec.rb
@@ -3,7 +3,7 @@ require 'rails_helper'
 RSpec.describe DefenceRequestsControllerPolicy do
   subject { DefenceRequestsControllerPolicy }
 
-  [:index?, :new?, :create?, :refresh_dashboard?].each do |action|
+  [:index?, :new?, :create?, :refresh_dashboard?, :edit?, :update?].each do |action|
     permissions action do
       it "grants access to #{action} user role is 'cso'" do
         expect(subject).to permit(User.new(role: :cso), DefenceRequestsController)
